Coerce point representation to int
unpackFloat returns float, but findPoint expects int. Therefore the
status quo ante never returned an actual point for me.

The floats I have seen in the wild are, in practice, very close to ints
(e.g. `1470.999755859375`), so Math.round() seems to convey the spirit
of the problem.

bmschmidt did *not* have this bug; I suspect it is very sensitive to
webgl implementation details within the browser and/or architecture.